refactor(tools): file splits, handler dedup, tree merge, exitInfo, nil-guard
Phase 2 Tasks 2.6+2.7+2.11 + Phase 3 Tasks 3.9+3.13+3.14 + Phase 4 Task 4.4:
- interface.go split: registry.go + registry_helpers.go (952→243+461+261)
- remote_sandbox.go split: _exec.go + _file.go (1384→1018+183+202)
- cd.go: formatDirectoryEntries shared core replaces twin buildDirectoryTree
- handler.go: handleFileOp[T] generic template replaces 8 copy-paste handlers
- exit_info.go: extractExitInfo unifies docker/native exitCode logic
- Dispatch: 9 bare strings → runnerproto.Proto* constants
- approval.go: nil-guard for isPermControlActiveFromCtx
