+++ tools/permission.go
@@ -12,3 +12,4 @@ func validateRunAsReason(runAs, reason string) error {
 	}
 	return nil
 }
+
