+++ tools/remote_sandbox.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"crypto/rand"
 	"crypto/subtle"
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io/fs"
@@ -694,178 +693,6 @@ func (rs *RemoteSandbox) GetShell(userID string, _ string) (string, error) {
 	return rc.shell, nil
 }
 
-func (rs *RemoteSandbox) Exec(ctx context.Context, spec ExecSpec) (*ExecResult, error) {
-	rc, err := rs.getRunner(spec.UserID)
-	if err != nil {
-		return nil, err
-	}
-
-	timeout := spec.Timeout
-	if timeout == 0 {
-		timeout = 60 * time.Second
-	}
-
-	reqBody, err := json.Marshal(ExecRequest{
-		Command: spec.Command,
-		Args:    spec.Args,
-		Shell:   spec.Shell,
-		Dir:     spec.Dir,
-		Env:     spec.Env,
-		Stdin:   spec.Stdin,
-		Timeout: int(timeout / time.Second),
-	})
-	if err != nil {
-		return nil, fmt.Errorf("marshal request: %w", err)
-	}
-
-	msg := &RunnerMessage{
-		ID:     generateID(),
-		Type:   ProtoExec,
-		UserID: spec.UserID,
-		Body:   reqBody,
-	}
-
-	resp, err := rs.sendRequest(ctx, rc, msg, timeout+5*time.Second)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.Type == ProtoError {
-		return nil, parseSandboxErrorResponse(resp.Body, "exec")
-	}
-
-	var result ExecResultResponse
-	if err := json.Unmarshal(resp.Body, &result); err != nil {
-		return nil, fmt.Errorf("unmarshal exec result: %w", err)
-	}
-
-	return &ExecResult{
-		Stdout:   result.Stdout,
-		Stderr:   result.Stderr,
-		ExitCode: result.ExitCode,
-		TimedOut: result.TimedOut,
-	}, nil
-}
-
-// ExecBg starts a background command on the runner.
-// Returns immediately with the task ID — the command runs asynchronously on the runner.
-func (rs *RemoteSandbox) ExecBg(ctx context.Context, spec ExecSpec, taskID string) error {
-	rc, err := rs.getRunner(spec.UserID)
-	if err != nil {
-		return err
-	}
-
-	reqBody, err := json.Marshal(runnerproto.BgExecRequest{
-		TaskID:  taskID,
-		Command: spec.Command,
-		Args:    spec.Args,
-		Shell:   spec.Shell,
-		Dir:     spec.Dir,
-		Env:     spec.Env,
-		Stdin:   spec.Stdin,
-	})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-
-	msg := &RunnerMessage{
-		ID:     generateID(),
-		Type:   runnerproto.ProtoBgExec,
-		UserID: spec.UserID,
-		Body:   reqBody,
-	}
-
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return err
-	}
-
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "bg_exec")
-	}
-
-	return nil
-}
-
-// KillBg kills a background task on the runner.
-func (rs *RemoteSandbox) KillBg(ctx context.Context, userID, taskID string) error {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return err
-	}
-
-	reqBody, err := json.Marshal(runnerproto.BgKillRequest{TaskID: taskID})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{
-		ID:     generateID(),
-		Type:   runnerproto.ProtoBgKill,
-		UserID: userID,
-		Body:   reqBody,
-	}
-
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return err
-	}
-
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "bg_kill")
-	}
-
-	return nil
-}
-
-// StatusBg queries the current status and output of a background task on the runner.
-func (rs *RemoteSandbox) StatusBg(ctx context.Context, userID, taskID string) (*RemoteBgTaskStatus, error) {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return nil, err
-	}
-
-	reqBody, err := json.Marshal(runnerproto.BgStatusRequest{TaskID: taskID})
-	if err != nil {
-		return nil, fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{
-		ID:     generateID(),
-		Type:   runnerproto.ProtoBgStatus,
-		UserID: userID,
-		Body:   reqBody,
-	}
-
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.Type == ProtoError {
-		return nil, parseSandboxErrorResponse(resp.Body, "bg_status")
-	}
-
-	var result runnerproto.BgOutputResponse
-	if err := json.Unmarshal(resp.Body, &result); err != nil {
-		return nil, fmt.Errorf("unmarshal bg_status result: %w", err)
-	}
-
-	return &RemoteBgTaskStatus{
-		TaskID:   result.TaskID,
-		Status:   result.Status,
-		ExitCode: result.ExitCode,
-		Stdout:   result.Stdout,
-		Stderr:   result.Stderr,
-	}, nil
-}
-
-// RemoteBgTaskStatus holds the status snapshot of a remote background task.
-type RemoteBgTaskStatus struct {
-	TaskID   string
-	Status   string // "running", "completed", "failed", "killed"
-	ExitCode int
-	Stdout   string
-	Stderr   string
-}
 
 // LLMGenerate sends an LLM generation request to the runner and returns the response.
 // This is used by ProxyLLM to forward LLM calls to runners with local LLM configured.
@@ -939,198 +766,6 @@ func (rs *RemoteSandbox) LLMModels(ctx context.Context, userID string) ([]string
 	return result.Models, nil
 }
 
-func (rs *RemoteSandbox) ReadFile(ctx context.Context, path, userID string) ([]byte, error) {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return nil, err
-	}
-
-	reqBody, err := json.Marshal(ReadFileRequest{Path: path})
-	if err != nil {
-		return nil, fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoReadFile, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return nil, err
-	}
-	if resp.Type == ProtoError {
-		return nil, parseSandboxErrorResponse(resp.Body, "read file")
-	}
-	var fc FileContentResponse
-	if err := json.Unmarshal(resp.Body, &fc); err != nil {
-		return nil, fmt.Errorf("unmarshal file content: %w", err)
-	}
-	return base64.StdEncoding.DecodeString(fc.Data)
-}
-
-func (rs *RemoteSandbox) WriteFile(ctx context.Context, path string, data []byte, perm os.FileMode, userID string) error {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return err
-	}
-
-	reqBody, err := json.Marshal(WriteFileRequest{
-		Path: path,
-		Data: base64.StdEncoding.EncodeToString(data),
-		Perm: int(perm),
-	})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoWriteFile, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return err
-	}
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "write file")
-	}
-	return nil
-}
-
-func (rs *RemoteSandbox) Stat(ctx context.Context, path, userID string) (*SandboxFileInfo, error) {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return nil, err
-	}
-	reqBody, err := json.Marshal(StatRequest{Path: path})
-	if err != nil {
-		return nil, fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoStat, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return nil, err
-	}
-	if resp.Type == ProtoError {
-		return nil, parseSandboxErrorResponse(resp.Body, "stat")
-	}
-	var sr StatResponse
-	if err := json.Unmarshal(resp.Body, &sr); err != nil {
-		return nil, fmt.Errorf("unmarshal stat: %w", err)
-	}
-	modTime, _ := time.Parse(time.RFC3339, sr.ModTime)
-	return &SandboxFileInfo{
-		Name:    sr.Name,
-		Size:    sr.Size,
-		Mode:    os.FileMode(sr.Mode),
-		ModTime: modTime,
-		IsDir:   sr.IsDir,
-	}, nil
-}
-
-func (rs *RemoteSandbox) ReadDir(ctx context.Context, path, userID string) ([]DirEntry, error) {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return nil, err
-	}
-	reqBody, err := json.Marshal(ReadDirRequest{Path: path})
-	if err != nil {
-		return nil, fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoReadDir, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return nil, err
-	}
-	if resp.Type == ProtoError {
-		return nil, parseSandboxErrorResponse(resp.Body, "read_dir")
-	}
-	var de DirEntriesResponse
-	if err := json.Unmarshal(resp.Body, &de); err != nil {
-		return nil, fmt.Errorf("unmarshal dir entries: %w", err)
-	}
-	entries := make([]DirEntry, len(de.Entries))
-	for i, e := range de.Entries {
-		entries[i] = DirEntry{Name: e.Name, IsDir: e.IsDir, Size: e.Size} //nolint:staticcheck
-	}
-	return entries, nil
-}
-
-func (rs *RemoteSandbox) MkdirAll(ctx context.Context, path string, perm os.FileMode, userID string) error {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return err
-	}
-	reqBody, err := json.Marshal(PathRequest{Path: path, Perm: int(perm)})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoMkdirAll, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return err
-	}
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "mkdir_all")
-	}
-	return nil
-}
-
-func (rs *RemoteSandbox) Remove(ctx context.Context, path, userID string) error {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return err
-	}
-	reqBody, err := json.Marshal(PathRequest{Path: path})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoRemove, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return err
-	}
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "remove")
-	}
-	return nil
-}
-
-func (rs *RemoteSandbox) RemoveAll(ctx context.Context, path, userID string) error {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return err
-	}
-	reqBody, err := json.Marshal(PathRequest{Path: path})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoRemoveAll, UserID: userID, Body: reqBody}
-	resp, err := rs.sendRequest(ctx, rc, msg, defaultRequestTimeout)
-	if err != nil {
-		return err
-	}
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "remove_all")
-	}
-	return nil
-}
-
-func (rs *RemoteSandbox) DownloadFile(ctx context.Context, url, outputPath, userID string) error {
-	rc, err := rs.getRunner(userID)
-	if err != nil {
-		return err
-	}
-	reqBody, err := json.Marshal(DownloadFileRequest{
-		URL:        url,
-		OutputPath: outputPath,
-	})
-	if err != nil {
-		return fmt.Errorf("marshal request: %w", err)
-	}
-	msg := &RunnerMessage{ID: generateID(), Type: ProtoDownloadFile, UserID: userID, Body: reqBody}
-	// 5-minute timeout for downloads
-	resp, err := rs.sendRequest(ctx, rc, msg, 5*time.Minute)
-	if err != nil {
-		return err
-	}
-	if resp.Type == ProtoError {
-		return parseSandboxErrorResponse(resp.Body, "download_file")
-	}
-	return nil
-}
 
 // === Runner sync (server → runner file sync on registration) ===
